fix: WeatherInfoListView 생성
