+++ WeatherForecast/WeatherForecast/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         guard let scene = (scene as? UIWindowScene) else { return }
         
-        let viewController: WeatherInfoListVC = WeatherInfoListVC(fetchDataManager: FetchDataManager(), imageManager: ImageManager())
+        let viewController: WeatherInfoListVC = WeatherInfoListVC()
         let navigationController: UINavigationController = UINavigationController(rootViewController: viewController)
 
         let window: UIWindow = UIWindow(windowScene: scene)
