+++ WeatherForecast/WeatherForecast/WeatherJSON/Weather.swift
@@ -5,7 +5,8 @@
 // 
 
 import Foundation
-
+// change structs to classes (video 26:00)
+// Or change reference types to enums (video 27:20)
 // MARK: - Weather JSON Format
 struct WeatherJSON: Decodable {
     let weatherForecast: [WeatherForecastInfo]
@@ -14,7 +15,7 @@ struct WeatherJSON: Decodable {
 
 // MARK: - List
 struct WeatherForecastInfo: Decodable {
-    let dt: TimeInterval
+    let dt: TimeInterval // Double
     let main: MainInfo
     let weather: Weather
     let dtTxt: String
@@ -49,14 +50,4 @@ struct Coord: Decodable {
     let lat, lon: Double
 }
 
-// MARK: - Temperature Unit
-enum TempUnit: String {
-    case metric, imperial
-    var expression: String {
-        switch self {
-        case .metric: return "℃"
-        case .imperial: return "℉"
-        }
-    }
-}
